Colours - types/palette
1. moved Colours type into colours file
2. added colours/changed names to be clearer
